implemented react contexts
